contentEntries as a function
